@@ -16,21 +16,21 @@
         }
     },
     "require": {
-        "php": ">=7.1",
+        "php": ">=8.3",
         "ext-json": "*",
         "gupalo/json": "*",
-        "promphp/prometheus_client_php": "^1.0|^2.11",
-        "symfony/http-foundation": "^3.0|^4.0|^5.0|^6.0|^7.0"
+        "promphp/prometheus_client_php": "^2.11",
+        "symfony/http-foundation": "^6.4|^7.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.5|^8.5|^9.5|^10.5"
+        "phpunit/phpunit": "^10.5"
     },
     "config": {
         "preferred-install": {
             "*": "dist"
         },
         "platform": {
-            "php": "7.1"
+            "php": "8.3"
         },
         "sort-packages": true
     }
