@@ -11,7 +11,7 @@ class JsonAdapter extends InMemory implements JsonSerializable
         array $counters = [],
         array $gauges = [],
         array $histograms = [],
-        array $summaries = []
+        array $summaries = [],
     ) {
         $this->counters = $counters;
         $this->gauges = $gauges;
@@ -29,13 +29,18 @@ class JsonAdapter extends InMemory implements JsonSerializable
         ];
     }
 
-    public static function createFromJson(array $data): self
+    public static function createFromJson(array $data): static
     {
-        return new self(
-            $data['counters'] ?? [],
-            $data['gauges'] ?? [],
-            $data['histograms'] ?? [],
-            $data['summaries'] ?? []
+        return static::create($data);
+    }
+
+    public static function create(array $data): static
+    {
+        return new static(
+            counters: $data['counters'] ?? [],
+            gauges: $data['gauges'] ?? [],
+            histograms: $data['histograms'] ?? [],
+            summaries: $data['summaries'] ?? [],
         );
     }
 }
